update 클라 서버
클라 점수표 반영 추가
클라 좌표조정 함수 조정
클라 셋타이머 개선
클라 라켓이미지 출력 개선

+++ AirHockey/AirHockey/framework.h
@@ -24,8 +24,8 @@
 //플레이어 에 대한 변수와 구조체 선언
 struct Point2D
 {
-    int Position_x;
-    int Position_y;
+    float Position_x;
+    float Position_y;
 };
 
 struct Accel2D
